Updates required properties typeddict to use schema_io template

+++ samples/client/petstore/python/src/petstore_api/components/schema/banana_req.py
@@ -24,8 +24,8 @@ BananaReqRequiredDictInput = typing_extensions.TypedDict(
     'BananaReqRequiredDictInput',
     {
         "lengthCm": typing.Union[
-            float,
-            int
+            int,
+            float
         ],
     }
 )
